requirements.txt
# requirements.txt
streamlit>=1.29.0
psycopg2-binary>=2.9.9
pandas>=2.1.4
plotly>=5.18.0
sqlparse>=0.4.4
redis>=5.0.1
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0